+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added new `Mastermind` archetype (Isekai Protagonist).
 - Added new `Overlord` archetype (Isekai Protagonist).
 - Added features to `Isekai Protagonist` progression: `Gifted`, `Release Energy`, `Afterimage`, `Secret Power`, `Hax`.
-- Added features to `Edge Lord` progression: `Chuunibyou Power`.
+- Added features to `Edge Lord` progression: `Chuunibyou Actualisation`, `Hidden Power`.
 - Added features to `God Emperor` progression: `Energy Condensation`, `Energy Barrier`, `Alteration of body and mind`, `Chosen Path`, `Transcendental Realm`.
 - Added features to `Hero` progression: `Hands of Salvation`, `Gold Barrier`, `Improved Gold Barrier`, `Greater Gold Barrier`, `Grand Gold Barrier`, `Deus Ex Machina`.
 - Added `Divine Aura` feature to `Otherwordly Aura` selection.
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 	- Updated icon for `Starting Weapon` feature.
 - Changes for `Edge Lord`:
 	- Now has Exotic Weapon Proficiency at 1st level.
+	- Added `Hidden Power` at 3rd level.
 	- Added another OP ability at 10th level.
 	- Added `Otherwordly Aura` at 10th level.
 - Changes for `God Emperor`: